requirements-api.txt
# FastAPI + server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.3.0

# Settings
pydantic-settings>=2.0.0

# ML inference
torch>=2.1.0
torchvision>=0.16.0
timm>=1.0.0
Pillow>=10.0.0

# Prototype computation
scikit-learn>=1.4.0
numpy>=1.26.0

# Config
pyyaml>=6.0

# Migration script
pandas>=2.2.0
openpyxl>=3.1.0
